car_reviews_project/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2==2.9.6
openpyxl==3.1.2
pandas==2.0.3
djangorestframework-simplejwt==5.2.2
